proper placement of static vars
it makes no sense to have static declarations in a header -
each compilation unit will have its own private version
of a static variable. if 'verbose' was meant to be changed
from client code, there should have been a function to set it.
the other variable was meant to be private, judging by the
comment, so wtf?